alchemy/cloud_function/inference/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage
google-cloud-secret-manager
envparse==0.2.0
